requirements.txt
nltk==3.8.1
openai==1.12.0
python-dotenv==1.0.1
kagglehub[pandas-datasets]==0.1.0
pandas==2.2.0
requests==2.32.3
selenium==4.18.1
webdriver-manager==4.0.1
beautifulsoup4==4.12.3 